Java runtime: fan.sys.Float => java.lang.Double

+++ src/jfan/fan/sys/Buf.java
@@ -234,9 +234,9 @@ public abstract class Buf
 
   public final Buf writeI8(Int x) { out.writeI8(x); return this; }
 
-  public final Buf writeF4(Float x) { out.writeF4(x); return this; }
+  public final Buf writeF4(Double x) { out.writeF4(x); return this; }
 
-  public final Buf writeF8(Float x) { out.writeF8(x); return this; }
+  public final Buf writeF8(Double x) { out.writeF8(x); return this; }
 
   public final Buf writeDecimal(Decimal x) { out.writeDecimal(x); return this; }
 
@@ -290,9 +290,9 @@ public abstract class Buf
 
   public final Int readS8() { return in.readS8(); }
 
-  public final Float readF4() { return in.readF4(); }
+  public final Double readF4() { return in.readF4(); }
 
-  public final Float readF8() { return in.readF8(); }
+  public final Double readF8() { return in.readF8(); }
 
   public final Decimal readDecimal() { return in.readDecimal(); }
